requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
python-socketio==5.10.0
python-engineio==4.8.1
eventlet==0.33.3
python-dotenv==1.0.0
gunicorn==21.2.0
